fixed licensing issues so that build succeeds and refactored some code
