+++ pinot-tools/src/main/java/com/linkedin/pinot/tools/SchemaInfo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.linkedin.pinot.tools.admin;
+package com.linkedin.pinot.tools;
 
 import java.util.Arrays;
 import java.util.List;
